product_engine/src/requirements.txt
fastapi==0.104.1
pytest==7.4.3
uvicorn==0.24.0
asyncpg==0.29.0
sqlalchemy==2.0.23
pydantic==2.5.2
requests==2.31.0
APScheduler==3.10.4
aiokafka==0.10.0
httpx==0.27.0
pydantic-settings==2.2.1
numpy-financial==1.0.0
python-dateutil==2.8.1